leetcode/54-spiral-matrix/des1/main_test.go
package des1

import (
	"github.com/stretchr/testify/require"
	"testing"
)

func Test_spiralOrder(t *testing.T) {
	type args struct {
		matrix [][]int
	}
	tests := []struct {
		name string
		args args
		want []int
	}{
		{
			name: "case1",
			args: args{
				matrix: [][]int{
					[]int{1, 2, 3},
					[]int{4, 5, 6},
					[]int{7, 8, 9},
				},
			},
			want: []int{1, 2, 3, 6, 9, 8, 7, 4, 5},
		},
		{
			name: "case2",
			args: args{
				matrix: [][]int{
					[]int{1, 2, 3, 4},
					[]int{5, 6, 7, 8},
					[]int{9, 10, 11, 12},
				},
			},
			want: []int{1, 2, 3, 4, 8, 12, 11, 10, 9, 5, 6, 7},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			got := spiralOrder(tt.args.matrix)
			require.Equal(t, tt.want, got)
		})
	}
}
